requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=4.5=1_gnu
absl-py=1.0.0=pypi_0
albumentations=0.4.6=pypi_0
blas=1.0=mkl
bottleneck=1.3.4=py38hce1f21e_0
ca-certificates=2022.4.26=h06a4308_0
cachetools=5.0.0=pypi_0
certifi=2021.10.8=py38h06a4308_2
charset-normalizer=2.0.12=pypi_0
cycler=0.11.0=pypi_0
einops=0.4.1=pypi_0
fonttools=4.31.2=pypi_0
google-auth=2.6.2=pypi_0
google-auth-oauthlib=0.4.6=pypi_0
grpcio=1.44.0=pypi_0
idna=3.3=pypi_0
imageio=2.16.1=pypi_0
imgaug=0.4.0=pypi_0
importlib-metadata=4.11.3=pypi_0
intel-openmp=2021.4.0=h06a4308_3561
joblib=1.1.0=pypi_0
kiwisolver=1.4.2=pypi_0
ld_impl_linux-64=2.35.1=h7274673_9
libffi=3.3=he6710b0_2
libgcc-ng=9.3.0=h5101ec6_17
libgomp=9.3.0=h5101ec6_17
libstdcxx-ng=9.3.0=hd4cf53a_17
markdown=3.3.6=pypi_0
matplotlib=3.5.1=pypi_0
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py38h7f8727e_0
mkl_fft=1.3.1=py38hd3c417c_0
mkl_random=1.2.2=py38h51133e4_0
ncurses=6.3=h7f8727e_2
networkx=2.7.1=pypi_0
numexpr=2.8.1=py38h6abb31d_0
numpy=1.22.3=pypi_0
numpy-base=1.21.5=py38hf524024_2
oauthlib=3.2.0=pypi_0
opencv-python=4.5.5.64=pypi_0
openssl=1.1.1n=h7f8727e_0
packaging=21.3=pyhd3eb1b0_0
pandas=1.4.2=py38h295c915_0
pillow=9.1.0=pypi_0
pip=21.2.4=py38h06a4308_0
protobuf=3.20.0=pypi_0
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pyparsing=3.0.7=pypi_0
python=3.8.13=h12debd9_0
python-dateutil=2.8.2=pyhd3eb1b0_0
pytorchcv=0.0.67=pypi_0
pytz=2021.3=pyhd3eb1b0_0
pywavelets=1.3.0=pypi_0
pyyaml=6.0=pypi_0
readline=8.1.2=h7f8727e_1
requests=2.27.1=pypi_0
requests-oauthlib=1.3.1=pypi_0
rsa=4.8=pypi_0
scikit-image=0.19.2=pypi_0
scikit-learn=1.0.2=pypi_0
scipy=1.8.0=pypi_0
setuptools=58.0.4=py38h06a4308_0
shapely=1.8.1.post1=pypi_0
six=1.16.0=pyhd3eb1b0_1
sqlite=3.38.2=hc218d9a_0
tensorboard=2.8.0=pypi_0
tensorboard-data-server=0.6.1=pypi_0
tensorboard-plugin-wit=1.8.1=pypi_0
threadpoolctl=3.1.0=pypi_0
tifffile=2022.3.25=pypi_0
tk=8.6.11=h1ccaba5_0
torch=1.8.1+cu102=pypi_0
torchaudio=0.8.0=pypi_0
torchsummary=1.5.1=pypi_0
torchvision=0.9.1+cu102=pypi_0
tqdm=4.64.0=pypi_0
typing-extensions=4.1.1=pypi_0
urllib3=1.26.9=pypi_0
werkzeug=2.1.1=pypi_0
wheel=0.37.1=pyhd3eb1b0_0
xz=5.2.5=h7b6447c_0
zipp=3.8.0=pypi_0
zlib=1.2.11=h7f8727e_4
